Rename the js/src/utils/recordEvent.js file to tracks.js to reduce confusion.
