+++ js/src/product-feed/issues-table-card/issues-table.js
@@ -10,7 +10,7 @@ import { Pagination, TablePlaceholder } from '@woocommerce/components';
  */
 import { ISSUE_TABLE_PER_PAGE } from '.~/constants';
 import ISSUES_TABLE_DATA_HEADERS from './issues-table-data-headers';
-import { recordTablePageEvent } from '.~/utils/recordEvent';
+import { recordTablePageEvent } from '.~/utils/tracks';
 import IssuesTableData from './issues-table-data';
 import useMCIssuesTypeFilter from '.~/hooks/useMCIssuesTypeFilter';
 import usePagination from '.~/hooks/usePagination';
